cartoonifyImage-main/cartoonifyImage-main/requirements.txt


opencv-python==4.3.0.36
streamlit==0.63.0
Pillow==7.0.0
numpy==1.18.1
